pipeline/requirements.txt
boto3
openpyxl
pandas
redshift_connector
python-dotenv
pymysql
ipykernel
altair
cryptography
awswrangler
streamlit
pyathena